zmf4: it seems that objects are versioned
Reasoning: If an image is saved with a non-default placement, the object
is 4 bytes longer and the "version" word of the object's header changes
from 0 to 1. 1 is also used in file header to mark files created by Draw
5. User guide to Draw 4 does not mention image placement, so I assume it
was added in v.5 (I don't have v.4 available right now, so I can't
check).